Updated mlx, improved rendering, fixed issue with saving to bmp

+++ includes/cub3d.h
@@ -6,7 +6,7 @@
 /*   By: bmans <marvin@codam.nl>                      +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2020/02/28 14:33:23 by bmans         #+#    #+#                 */
-/*   Updated: 2020/06/30 16:53:27 by brendan       ########   odam.nl         */
+/*   Updated: 2020/07/01 15:56:20 by bmans         ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -121,7 +121,7 @@ void			vec2_rot(double *out, double *in, double r, double theta);
 void			prepare_window(t_world *world, const char *name);
 void			render(t_world *world, char to_window);
 u_int32_t		*get_pixel(t_texture *tex, int *pixel);
-u_int32_t		*get_pixel_scaled(t_texture *tex, double *point);
+u_int32_t		*get_pixel_scaled(t_texture *tex, float *point);
 void			render_floor_ceiling(t_world *world);
 char			check_map(char **map);
